Moves check for update type to Event component
Change-type: patch
