+++ lib/Event/index.jsx
@@ -8,7 +8,6 @@ import React from 'react'
 import {
 	compose
 } from 'redux'
-import Event from './Event'
 import {
 	getMessage
 } from './EventBody'
@@ -19,18 +18,38 @@ import {
 	withSetup
 } from '../SetupProvider'
 
+import {
+	UPDATE
+} from '../constants'
+import Update from '../Update'
+import Event from './Event'
+
 export {
 	getMessage
 }
 
 const EventWithActor = (props) => {
+	const {
+		card, user, selectCard, getCard, getActor, onCardVisible
+	} = props
+	const typeBase = props.card.type.split('@')[0]
+	if (typeBase === UPDATE) {
+		return (
+			<Update
+				onCardVisible={onCardVisible}
+				card={card}
+				user={user}
+				getActor={getActor}
+			/>
+		)
+	}
 	return (
 		<CardLoader
 			id={helpers.getActorIdFromCard(props.card)}
 			type="user"
 			withLinks={[ 'is member of' ]}
-			cardSelector={props.selectCard}
-			getCard={props.getCard}
+			cardSelector={selectCard}
+			getCard={getCard}
 		>
 			{(author) => {
 				return (
